split provider impls into separate .scala files
